Combine classes into module
